# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"CommentResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"ideaId": {
						"type": "string"
					},
					"parentCommentId": {
						"type": "string",
						"nullable": true
					},
					"authorId": {
						"type": "string"
					},
					"author": {
						"properties": {
							"icon": {
								"type": "string",
								"nullable": true
							},
							"displayName": {
								"type": "string",
								"nullable": true
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"_count": {
						"properties": {
							"replies": {
								"type": "number",
								"format": "double"
							},
							"likes": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"replies",
							"likes"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"content",
					"ideaId",
					"authorId",
					"author",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetCommentsResponse": {
				"properties": {
					"comments": {
						"items": {
							"$ref": "#/components/schemas/CommentResponse"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"hasPrev": {
								"type": "boolean"
							},
							"hasNext": {
								"type": "boolean"
							},
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"hasPrev",
							"hasNext",
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"comments",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCommentInput": {
				"properties": {
					"ideaId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"parentCommentId": {
						"type": "string"
					}
				},
				"required": [
					"ideaId",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SigninOutput": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					},
					"user": {
						"properties": {
							"updated_at": {
								"type": "string"
							},
							"created_at": {
								"type": "string"
							},
							"userProfile": {
								"properties": {
									"followersCount": {
										"type": "number",
										"format": "double"
									},
									"followingCount": {
										"type": "number",
										"format": "double"
									},
									"links": {
										"items": {
											"properties": {
												"url": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"url",
												"id"
											],
											"type": "object"
										},
										"type": "array"
									},
									"backgroundImage": {
										"type": "string",
										"nullable": true
									},
									"icon": {
										"type": "string",
										"nullable": true
									},
									"autobiography": {
										"type": "string",
										"nullable": true
									},
									"displayName": {
										"type": "string",
										"nullable": true
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"followersCount",
									"followingCount",
									"links",
									"backgroundImage",
									"icon",
									"autobiography",
									"displayName",
									"id"
								],
								"type": "object",
								"nullable": true
							},
							"verified": {
								"type": "boolean"
							},
							"email": {
								"type": "string"
							},
							"username": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"updated_at",
							"created_at",
							"userProfile",
							"verified",
							"email",
							"username",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"accessToken",
					"refreshToken",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
...
		"/v1/comment": {
			"get": {
				"operationId": "GetComments",
				"responses": {
					"200": {
						"description": "Comments retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/GetCommentsResponse"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Get comments for an idea with pagination\n\nThis endpoint allows authenticated users to retrieve comments for a specific idea.\nComments are returned in a paginated format and can be filtered by parent comment.",
				"tags": [
					"Comment"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the idea to get comments for.",
						"in": "query",
						"name": "ideaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Optional page number for pagination (default: 1).",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Optional number of items per page (default: 10, max: 100).",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Optional ID of parent comment to get replies (for nested comments).",
						"in": "query",
						"name": "parentCommentId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"201": {
						"description": "Comment created successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Create a new comment on an idea\n\nThis endpoint allows authenticated users to create comments on ideas or reply\nto existing comments. Comments support nested replies through parentCommentId.\n\nComment Creation Process:\n- Validates user authentication\n- Validates that the target idea exists\n- Validates parent comment if replying to a comment\n- Creates comment in database with author information\n- Returns created comment with engagement metrics",
				"tags": [
					"Comment"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Object containing ideaId, content, and optional parentCommentId",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentInput",
								"description": "- Object containing ideaId, content, and optional parentCommentId"
							}
						}
					}
				}
			}
		},
		"/v1/auth/callback/{provider}": {
			"get": {
				"operationId": "CallbackSocialLoginGet",
				"responses": {
					"200": {
						"description": ""
					}
				},
				"description": "Handle OAuth callback from social providers\n\nThis endpoint processes the OAuth callback from social login providers (Google,\nLinkedIn, GitHub) after the user has authenticated with the external service.\nIt exchanges the authorization code for access tokens and completes the login process.\n\nOAuth Callback Process:\n- Receives authorization code from provider\n- Exchanges code for access tokens via AuthGateway\n- Retrieves user profile information\n- Creates or updates user account\n- Generates application authentication tokens",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- Social provider name (google, linkedin, github)",
						"in": "path",
						"name": "provider",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"google",
								"linkedin",
								"github"
							]
						}
					},
					{
						"description": "- Authorization code from OAuth provider",
						"in": "query",
						"name": "code",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "- State parameter for OAuth security",
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/social-login/{provider}": {
			"post": {
				"operationId": "SocialLogin",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SigninOutput"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Authenticate user with social provider token\n\nThis endpoint authenticates users using tokens obtained from social providers\n(Google, LinkedIn, GitHub). It validates the social token, retrieves user profile\ninformation, and either signs in existing users or creates new accounts.\n\nSocial Authentication Process:\n- Validates social provider token via AuthGateway\n- Retrieves user profile from social provider\n- Checks if user already exists by email\n- Creates new account or links to existing account\n- Generates application authentication tokens\n- Updates refresh token in database",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- Social provider name (google, linkedin, github)",
						"in": "path",
						"name": "provider",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"google",
								"linkedin",
								"github"
							]
						}
					}
				],
				"requestBody": {
					"description": "- Object containing the social provider token",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"token": {
										"type": "string"
									}
								},
								"required": [
									"token"
								],
								"type": "object",
								"description": "- Object containing the social provider token"
							}
						}
					}
				}
			}
		},
		"/v1/auth/link-new-oauth-provider/{provider}": {
			"post": {
				"operationId": "LinkNewOAuthProvider",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"success"
											],
											"type": "object"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Link additional OAuth provider to existing account\n\nThis endpoint allows authenticated users to link additional social providers\n(Google, LinkedIn, GitHub) to their existing account. This enables users to\nsign in with multiple social accounts using the same application account.\n\nProvider Linking Process:\n- Validates user authentication via JWT token\n- Validates social provider token via AuthGateway\n- Retrieves user profile from social provider\n- Checks if provider is already linked to another account\n- Links provider to current user account",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- Social provider name (google, linkedin, github)",
						"in": "path",
						"name": "provider",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"google",
								"linkedin",
								"github"
							]
						}
					}
				],
				"requestBody": {
					"description": "- Object containing the social provider token",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"token": {
										"type": "string"
									}
								},
								"required": [
									"token"
								],
								"type": "object",
								"description": "- Object containing the social provider token"
							}
						}
					}
				}
			}
		},
...
		"/v1/auth/refresh-token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"refreshToken": {
													"type": "string"
												},
												"accessToken": {
													"type": "string"
												}
											},
											"required": [
												"refreshToken",
												"accessToken"
											],
											"type": "object"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Refresh authentication tokens using refresh token\n\nThis endpoint allows users to obtain new access and refresh tokens using their\ncurrent refresh token. It validates the refresh token, verifies the user exists,\nand generates a new token pair. This is essential for maintaining user sessions\nwithout requiring re-authentication.\n\nToken Refresh Process:\n- Validates refresh token signature and expiration\n- Extracts user ID from token payload\n- Verifies user exists and is active\n- Generates new access and refresh tokens\n- Updates refresh token in database",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Object containing the refresh token",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"token": {
										"type": "string"
									}
								},
								"required": [
									"token"
								],
								"type": "object",
								"description": "- Object containing the refresh token"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}